refactor(ast/estree): split custom serializers into multiple files (#10967)
Pure refactor. `serialize.rs` had got too large. Split it up into multiple files.

Also make sure all custom serializers/converters have clear doc comments.
