+++ tasks/ast_tools/src/main.rs
@@ -212,7 +212,12 @@ static SOURCE_PATHS: &[&str] = &[
     "crates/oxc_ast/src/ast/jsx.rs",
     "crates/oxc_ast/src/ast/ts.rs",
     "crates/oxc_ast/src/ast/comment.rs",
-    "crates/oxc_ast/src/serialize.rs",
+    "crates/oxc_ast/src/serialize/mod.rs",
+    "crates/oxc_ast/src/serialize/basic.rs",
+    "crates/oxc_ast/src/serialize/literal.rs",
+    "crates/oxc_ast/src/serialize/js.rs",
+    "crates/oxc_ast/src/serialize/jsx.rs",
+    "crates/oxc_ast/src/serialize/ts.rs",
     "crates/oxc_syntax/src/lib.rs",
     "crates/oxc_syntax/src/module_record.rs",
     "crates/oxc_syntax/src/number.rs",
